+++ path/fount.sh
@@ -863,9 +863,9 @@ run() {
 		fi
 	fi
 	if [[ $# -gt 0 && $1 = 'debug' ]]; then
-		run_deno run --allow-scripts --allow-all --inspect-brk "$FOUNT_DIR/src/server/index.mjs" "${@:2}"
+		run_deno run --allow-scripts --allow-all --inspect-brk --node-modules-dir=none "$FOUNT_DIR/src/server/index.mjs" "${@:2}"
 	else
-		run_deno run --allow-scripts --allow-all "$FOUNT_DIR/src/server/index.mjs" "$@"
+		run_deno run --allow-scripts --allow-all --node-modules-dir=none "$FOUNT_DIR/src/server/index.mjs" "$@"
 	fi
 	local exit_code=$?
 	if [[ $IN_TERMUX -eq 1 ]]; then export LANG="$LANG_BACKUP"; fi
@@ -874,11 +874,14 @@ run() {
 
 # 安装 fount 依赖
 if [[ ! -d "$FOUNT_DIR/node_modules" || ($# -gt 0 && $1 = 'init') ]]; then
-	git -C "$FOUNT_DIR" clean -fd
+	if [ ! -f "$FOUNT_DIR/.noupdate" ]; then
+		git -C "$FOUNT_DIR" clean -fd
+		git -C "$FOUNT_DIR" reset --hard "origin/master"
+	fi
 	if [[ -d "$FOUNT_DIR/node_modules" ]]; then run "shutdown"; fi
 	echo "Installing dependencies..."
 	set +e # 禁用错误检测，因为第一次运行可能会失败
-	run_deno install --reload --allow-scripts --allow-all --node-modules-dir=auto --entrypoint "$FOUNT_DIR/src/server/index.mjs"
+	run_deno install --reload --allow-scripts --allow-all --node-modules-dir=none --entrypoint "$FOUNT_DIR/src/server/index.mjs"
 	run "shutdown" # 确保安装后服务能正常启动
 	set -e         # 重新启用严格模式
 	if [ $IN_DOCKER -eq 0 ] && [ $IN_TERMUX -eq 0 ]; then
